Updated the bank panel to handle more than 8 line items (by combining some together into an "Other" line).

+++ BankPanel.h
@@ -41,7 +41,7 @@ private:
 	
 private:
 	PlayerInfo &player;
-	int qualify;
+	int64_t qualify;
 	int selectedRow;
 };
